Add authentication check to author model

+++ server/resolvers/author.js
@@ -1,4 +1,6 @@
 const {AuthorModel} = require('../models/author');
+const {ForbiddenError} = require('apollo-server-express');
+
 
 module.exports.author = async (_, args, req) => {
     try {
@@ -28,39 +30,51 @@ module.exports.addAuthor = async (_, args, req) => {
 };
 
 module.exports.editAuthor = async (_, args, req) => {
-  try {
-      let author = await AuthorModel.findById(args.author._id);
-      if (!(author && (typeof author === 'object'))){
-          throw new Error("Author doesn't exist");
-      }
-      if(author.account.toString() !== args.author.account){
-          throw new Error("The account associated with the author object cannot be changed")
-      }
-      if(!args.author.posts && author.posts && Array.isArray(author.posts) && author.posts.length ){
-          args.author.posts = [];
-          author.posts.forEach(x => {
-              args.author.posts.push(x);
-          })
-      }
-      return await AuthorModel.findOneAndUpdate({_id: args.author._id}, {$set: args.author}, {new: true});
-  } catch (e){
-      console.log(e);
-  }
+    req.user = {...req.user, ...(await req.user.checkAuthentication())};
+    const reqUser = req.user._doc;
+    if(reqUser && reqUser._id && (reqUser._id.toString()=== args.author.account || reqUser.accountType === 'admin')) {
+        try {
+            let author = await AuthorModel.findById(args.author._id);
+            if (!(author && (typeof author === 'object'))) {
+                throw new Error("Author doesn't exist");
+            }
+            if (author.account.toString() !== args.author.account) {
+                throw new Error("The account associated with the author object cannot be changed")
+            }
+            if (!args.author.posts && author.posts && Array.isArray(author.posts) && author.posts.length) {
+                args.author.posts = [];
+                author.posts.forEach(x => {
+                    args.author.posts.push(x);
+                })
+            }
+            return await AuthorModel.findOneAndUpdate({_id: args.author._id}, {$set: args.author}, {new: true});
+        } catch (e) {
+            console.log(e);
+        }
+    } else {
+        throw new ForbiddenError('403-Forbidden');
+    }
 };
 
 module.exports.deleteAuthor = async (_, args, req) => {
-  try{
-      let author = await AuthorModel.findById(args.author._id);
-      if(author && (typeof author === 'object')){
-          if(args.author.account !== author.account.toString()){
-              throw new Error("The account _id doesn't match the author's account _id");
-          }
-          return await AuthorModel.findByIdAndDelete(args.author._id);
-      } else {
-          throw new Error("Couldn't find the author specified");
-      }
-  } catch (e) {
-      console.log(e);
-      throw new Error(e);
-  }
+    req.user = {...req.user, ...(await req.user.checkAuthentication())};
+    const reqUser = req.user._doc;
+    if(reqUser && reqUser._id && (reqUser._id.toString()=== args.author.account || reqUser.accountType === 'admin')) {
+        try {
+            let author = await AuthorModel.findById(args.author._id);
+            if (author && (typeof author === 'object')) {
+                if (args.author.account !== author.account.toString()) {
+                    throw new Error("The account _id doesn't match the author's account _id");
+                }
+                return await AuthorModel.findByIdAndDelete(args.author._id);
+            } else {
+                throw new Error("Couldn't find the author specified");
+            }
+        } catch (e) {
+            console.log(e);
+            throw new Error(e);
+        }
+    } else {
+        throw new ForbiddenError('403-Forbidden');
+    }
 };
